Clients completed

+++ server/routers/sales/packman.js
@@ -1,10 +1,10 @@
-const { Packman, validatePackman } = require('../../models/Sales/Packman.js');
-const { Market } = require('../../models/MarketAndBranch/Market');
+const { Packman, validatePackman } = require("../../models/Sales/Packman.js");
+const { Market } = require("../../models/MarketAndBranch/Market");
 
 module.exports.register = async (req, res) => {
   try {
     const { name, market, currentPage, countPage, search } = req.body;
-    const { error } = validatePackman({name, market});
+    const { error } = validatePackman({ name, market });
     if (error) {
       return res.status(400).json({
         error: error.message,
@@ -34,18 +34,24 @@ module.exports.register = async (req, res) => {
     });
 
     await newPackman.save();
-    const namepackman = new RegExp('.*' + search ? search.name : '' + '.*', 'i');
+    const namepackman = new RegExp(
+      ".*" + search ? search.name : "" + ".*",
+      "i"
+    );
 
-    const packmansCount = await Packman.find({ market, name: namepackman }).count();
+    const packmansCount = await Packman.find({
+      market,
+      name: namepackman,
+    }).count();
 
     const packmans = await Packman.find({ market, name: namepackman })
       .sort({ _id: -1 })
-      .select('name market')
+      .select("name market")
       .skip(currentPage * countPage)
       .limit(countPage);
     res.status(201).json({ packmans: packmans, count: packmansCount });
   } catch (error) {
-    res.status(400).json({ error: 'Serverda xatolik yuz berdi...' });
+    res.status(400).json({ error: "Serverda xatolik yuz berdi..." });
   }
 };
 
@@ -59,11 +65,10 @@ module.exports.getAll = async (req, res) => {
       });
     }
 
-    const packman = await Packman.find({ market }).select('name market');
-
+    const packman = await Packman.find({ market }).select("name market");
     res.status(201).send(packman);
   } catch (error) {
-    res.status(501).json({ error: 'Serverda xatolik yuz berdi...' });
+    res.status(501).json({ error: "Serverda xatolik yuz berdi..." });
   }
 };
 
@@ -86,10 +91,10 @@ module.exports.updatePackman = async (req, res) => {
     await Packman.findByIdAndUpdate(_id, {
       name: name,
     });
-    const updatePackman = await Packman.findById(_id).select('name market');
+    const updatePackman = await Packman.findById(_id).select("name market");
     res.status(201).send(updatePackman);
   } catch (error) {
-    res.status(501).json({ error: 'Serverda xatolik yuz berdi...' });
+    res.status(501).json({ error: "Serverda xatolik yuz berdi..." });
   }
 };
 
@@ -109,11 +114,17 @@ module.exports.deletePackman = async (req, res) => {
         .json({ message: `Diqqat! ${name} yetkazuvchi avval yaratilmagan!` });
     }
 
+    if (packman.clients && packman.clients.length > 0) {
+      return res.status(400).json({
+        message: `Diqqat!  Agentda mijozlar mavjudligi sababli agentni ro'yxatdan o'chirishning imkoni mavjud emas!`,
+      });
+    }
+
     await Packman.findByIdAndDelete(_id);
 
     res.status(201).send(packman);
   } catch (error) {
-    res.status(501).json({ error: 'Serverda xatolik yuz berdi...' });
+    res.status(501).json({ error: "Serverda xatolik yuz berdi..." });
   }
 };
 
@@ -127,17 +138,17 @@ module.exports.getPackmans = async (req, res) => {
         .json({ message: "Diqqat! Do'kon malumotlari topilmadi." });
     }
 
-    const name = new RegExp('.*' + search ? search.name : '' + '.*', 'i');
+    const name = new RegExp(".*" + search ? search.name : "" + ".*", "i");
 
     const packmansCount = await Packman.find({ market, name: name }).count();
 
     const packmans = await Packman.find({ market, name: name })
       .sort({ _id: -1 })
-      .select('name market')
+      .select("name market")
       .skip(currentPage * countPage)
       .limit(countPage);
     res.status(201).json({ packmans: packmans, count: packmansCount });
   } catch (error) {
-    res.status(501).json({ error: 'Serverda xatolik yuz berdi...' });
+    res.status(501).json({ error: "Serverda xatolik yuz berdi..." });
   }
 };
